+++ lib/bot_platform/dialogs.rb
@@ -4,11 +4,20 @@ require_relative 'dialogs/dialog_instance'
 require_relative 'dialogs/dialog_result'
 require_relative 'dialogs/dialog_set'
 require_relative 'dialogs/dialog_state'
+require_relative 'dialogs/dialog_reason'
 
 require_relative 'dialogs/prompts/prompt'
 require_relative 'dialogs/prompts/prompt_options'
+require_relative 'dialogs/prompts/confirm_prompt'
+require_relative 'dialogs/prompts/number_prompt'
 require_relative 'dialogs/prompts/text_prompt'
 require_relative 'dialogs/prompts/prompt_recognizer_result'
+
+
+require_relative 'dialogs/waterfall_dialog'
+require_relative 'dialogs/waterfall_step'
+require_relative 'dialogs/waterfall_step_context'
+
 
 module BotPlatform
   module Dialogs