+++ lib/bot_platform/dialogs/prompts/prompt_options.rb
@@ -6,6 +6,9 @@ module BotPlatform
       class PromptOptions
         attr_accessor :prompt, :retry_prompt, :choices, :validations
 
+        def initialize(prompt:nil)
+          @prompt = prompt
+        end
       end
     end
   end