+++ lib/bot_platform/state/bot_state.rb
@@ -11,6 +11,15 @@ module BotPlatform
         @service_key = service_key
       end
 
+      def create_property(key)
+        @storage.data[key.to_sym] = {}
+        return @storage.data[key.to_sym]
+      end
+
+      def get_property(key)
+        return @storage.data[key.to_sym]
+      end
+
       def save_changes(turn_context, force=false)
         assert_turn_context_is_valid turn_context