+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-09-06
+
+### Added
+- Waterfall dialog support
+- Waterfall sample
+- user_state
+
+### Fixed
+- end_of_turn response
+
 ## [0.2.2-0.2.3] - 2021-09-06
 
 ### Added